pyproject.toml
# >----------< PROJECT >----------<

[project]
name = "termdo_tasks_api"
version = "1.0.0"
description = "Manages task operations including create update delete and retrieval for user todos"
authors = [{ name = "Emrecan Karaçayır", email = "developer@karacayir.com" }]
license = "MIT"

[project.urls]
Repository = "https://github.com/termdo-inc/termdo-tasks-api"
Issues = "https://github.com/termdo-inc/termdo-tasks-api/issues"

# >----------< BUILD SYSTEM >----------<

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# >----------< POETRY >----------<

[tool.poetry]
requires-poetry = ">=2.0.0,<3.0.0"
packages = [{ include = "termdo_tasks_api", from = "source" }]

# >----------< DEPENDENCIES >----------<

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.30.0"
fastapi = "^0.116.1"
python-dotenv = "^1.1.1"
uvicorn = "^0.35.0"

[tool.poetry.group.dev.dependencies]
asyncpg-stubs = "^0.30.0"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

# >----------< CONFIGS >----------<

[tool.black]
line-length = 80

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = "tests/"
pythonpath = "source/"
addopts = ["--cov=source/", "--cov-report=term-missing"]
